chore: rename exports, change class namings, change mentions of basic for core (#4009)
