+++ apps/laboratory/src/components/AppKitCoreHooks.tsx
@@ -14,13 +14,13 @@ import {
   useAppKitAccount,
   useAppKitNetwork,
   useDisconnect
-} from '@reown/appkit/react-basic'
+} from '@reown/appkit/react-core'
 
-export interface AppKitBasicHooksProps {
+export interface AppKitCoreHooksProps {
   networks?: AppKitNetwork[]
 }
 
-export function AppKitBasicHooks({ networks }: AppKitBasicHooksProps) {
+export function AppKitCoreHooks({ networks }: AppKitCoreHooksProps) {
   const { open } = useAppKit()
   const { caipNetwork, switchNetwork } = useAppKitNetwork()
   const { caipAddress } = useAppKitAccount()
