+++ examples/html-ak-basic-up/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>HTML AppKit Basic Universal Provider Example</title>
+    <title>HTML AppKit Core Universal Provider Example</title>
     <link rel="stylesheet" href="/src/main.css" />
     <link
       rel="icon"
@@ -21,7 +21,7 @@
         <img src="/appkit-logo.png" alt="AppKit" width="150" />
       </div>
 
-      <h1 class="page-title">HTML AppKit Basic Universal Provider Example</h1>
+      <h1 class="page-title">HTML AppKit Core Universal Provider Example</h1>
 
       <div class="appkit-buttons-container">
         <div class="row">
