+++ examples/html-ak-basic-up/src/main.js
@@ -1,7 +1,7 @@
 import UniversalProvider from '@walletconnect/universal-provider'
 import base58 from 'bs58'
 
-import { createAppKit } from '@reown/appkit/basic'
+import { createAppKit } from '@reown/appkit/core'
 import { bitcoin, mainnet, polygon, solana } from '@reown/appkit/networks'
 
 // Constants
