+++ examples/vue-ak-basic-up/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vue AppKit Basic Universal Provider Example</title>
+    <title>Vue AppKit Core Universal Provider Example</title>
     <link
       rel="icon"
       type="image/png"
