+++ examples/vue-ak-basic/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vue AppKit Basic Example</title>
+    <title>Vue AppKit Core Example</title>
     <link
       rel="icon"
       type="image/png"
